Qt-removal R4.2 prereq: split ChatWorker/ChatAgent out of graphlink_agents_core
Discovered while scoping R4.2's real agent-dispatch service: R4.1 made
api_provider/graphlink_task_config importable with zero PySide6, but the
actual chat-completion classes the new backend needs - ChatWorker,
ChatAgent, resolve_branch_system_prompt - live in
graphlink_agents_core.py, which has its own unconditional
`from PySide6.QtCore import QThread, Signal, QPointF` (needed only by its
*WorkerThread classes). Importing anything from that module, including
these three Qt-free symbols, still pulled Qt - the R4.1 unblock didn't
reach this file.

Same fix shape as R4.1: extracted the three symbols verbatim into a new
Qt-free graphlink_chat_agent.py; graphlink_agents_core.py imports them
back for its own ChatWorkerThread and every legacy Qt call site (the
facade graphlink_agents.py re-exports through agents_core unchanged, so
no other file needed touching).

New proof: backend/tests/test_agent_layer_qt_free.py gains
test_chat_agent_imports_qt_free, a fresh-subprocess assertion that
import graphlink_chat_agent never pulls PySide6 into sys.modules.

1929 pytest green (up 1, the new test), zero behavior change, zero
regressions in any legacy Qt call site.
